Add the hook to create the services
The hook is implemented in Jsonnet. Kubernetes passes the StatefulSet into the function. The implementation iterates over each replica in the StatefulSet and builds a list of service attachments. Then we attach arbitrary resources that are only connected to the target resource through owner references, meaning Kubernetes will delete them if the StatefulSet is deleted.